# storage/api-docs/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "TasteAsia RESTful API",
        "description": "RESTful API Documentation for TasteAsia. Explore our endpoints and discover how to interact with our delicious data!",
        "contact": {
            "email": "gabrielaswinta@cosrsivalab.space"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://tasteasia.test",
            "description": "TasteAsia API Server"
        },
        {
            "url": "https://tasteasia.test/api/v1",
            "description": "Local Development Server"
        },
        {
            "url": "${STAGING_URL}/api/v1",
            "description": "STAGING API Server"
        },
        {
            "url": "https://tasteasia.test/api/v1",
            "description": "PRODUCTION API Server"
        }
    ],
    "paths": {
...
        "/m/wallet/topup": {
            "post": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Topup Cash Balance",
                "description": "Topup customer's cash balance with Hitpay Payment Gateway",
                "operationId": "7ca380f021ae73362608e2cf2a68a5d5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "description": "Amount to top-up",
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Top-up successful",
                                            "type": "string"
                                        },
                                        "status": {
                                            "description": "success",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    },
                    {
                        "Access Token": []
                    }
                ]
            }
        },
        "/m/wallet/history/{type}": {
            "post": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Usage history of cash or credit",
                "description": "Type of wallet history (cash or credit)",
                "operationId": "87cfecead19e6963b335c261a1786acc",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Type of wallet history (cash or credit)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Cash"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Top-up successful",
                                            "type": "string"
                                        },
                                        "status": {
                                            "description": "success",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    },
                    {
                        "Access Token": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "Bearer Token": {
                "type": "http",
                "description": "Enter token in format 'Bearer {token}'. Token is a 64-character random string.",
                "scheme": "bearer"
            },
            "Access Token": {
                "type": "apiKey",
                "description": "Sanctum Access Token (a random string)",
                "name": "X-Taste-Asia-Access-Token",
                "in": "header"
            },
            "Refresh Token": {
                "type": "apiKey",
                "description": "Sanctum Refresh Token (a random string)",
                "name": "X-Taste-Asia-Refresh-Token",
                "in": "header"
            }
        }
    },
...
}